Trim unrelated E2E changes from leaderboard PR
